+++ src/methods/$group-by/async-group-by.d.ts
@@ -7,16 +7,9 @@
  */
 
 import { AsyncSourceIterable, AsyncResultIterable } from '../../types/async-iterable';
-declare function asyncGroupBy(
-  key: null | undefined,
-): <T>(source: AsyncSourceIterable<T>) => AsyncResultIterable<[T, AsyncResultIterable<T>]>;
 declare function asyncGroupBy<K, T>(
   key: (item: T) => K | Promise<K>,
 ): (source: AsyncSourceIterable<T>) => AsyncResultIterable<[K, AsyncResultIterable<T>]>;
-declare function asyncGroupBy<T>(
-  key: null | undefined,
-  source: AsyncSourceIterable<T>,
-): AsyncResultIterable<[T, AsyncResultIterable<T>]>;
 declare function asyncGroupBy<K, T>(
   key: (item: T) => K | Promise<K>,
   source: AsyncSourceIterable<T>,